[operators] introduce 'solver_options'
- 'options' is removed from OperatorInterface.apply_inverse
- 'invert_options' is removed from OperatorInterface
- 'solver_options' is added to OperatorInterface

If not 'None', solver_options is a dict with configurations for the
linear (and in future nonlinear, least-squares) solvers of the various
opertor types in pyMOR. Currently, only the keys 'numpy_dense',
'numpy_sparse' and 'generic' are recognized.

As a rule of thumb, new operators obtained from a given operator
(projection, Jacobian, linear combination) inherit its solver_options.
Each Operator implementation should take 'solver_options' as an __init__
argument in order to being able to customize the solver options for
different operators.