[demos] update thermalblock demos to use h1_0_semi_product
h1_(semi)_product cannot be used, as it will yield wrong result when
used to calculate Riesz representatives.

Note that only for 'error_product' the use of h1_0_semi_product
instead of h1_semi_product is mandatory. For norm calculation, both
product give the same result for H1_0 functions.

We use again h1_0_semi_product instead of h1_0_product, as the former is
a more natural norm for the problem.