general refactoring.
Rolling back to the classic pyMOR Riccati interface without Q. Instead now the required Q in RADI is available only in the `lrradi` worker routine called by the `solve_ricc_lrcf`.

No rewriting of test infrastructure required and LQG-BT, PR-BT and BR-BT actually do not have a Q in the first place.  So, once `lrradi` converges properly, this should be the least intrusive way to implement thing.

However, on a simple test case from the pyMOR tests it currently uses about 5x the iterations compared to the MESS implementation and still only achieves half the accuracy in number of correct digits.