Adjusted tolerances for tests
Reverted changes from previous commit and increased the `check_orth` tolerance for `cgs_iro_ls` instead.
For too high matrix condition numbers the `Q` has a too higher loss of orthogonality, which causes tolerance errors in the test.
Now an `AccuracyError` should be thrown, which would check for the matrix condition number and fail or pass the test accordingly.