fixed a bug in the neural_network reductor. When a multiple_restarts_training is called it will first run the #0 training and check if the error is below the target loss before it then starts multiple other restarts to check if any of these reach below the target_loss.
Right now the neural_network of the last epoch is returned when the training stops after the #0 run. Instead it should be the best_neural network as in line 1204.
This bug showed when I trained a neural network with very limited data. It ran fine in the first epochs, but then the loss exploded (1e+24), Early stopping was called and I got back the neural network architecture with a huge error on the data.